backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pillow==10.4.0
psd-tools==1.5.0
imageio==2.35.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.2
